requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
RPi.GPIO==0.7.1
spidev==3.7
